feat : ajout list categories et auteurs

+++ bookhub-back/src/main/java/fr/eni/bookhub/config/ModelMapperConfig.java
@@ -22,7 +22,7 @@ public class ModelMapperConfig {
         modelMapper.typeMap(CreateLivreDTO.class, Livre.class).addMappings(mapper -> {
             mapper.skip(Livre::setId);
             mapper.map(CreateLivreDTO::getAuteurId, (dest, v) -> dest.getAuteur().setId((Integer) v));
-            mapper.map(CreateLivreDTO::getCategorieId, (dest, v) -> dest.getCategorie().setId((Integer) v));
+            mapper.map(CreateLivreDTO::getCategorieId, (dest, v) -> dest.getCategorie().setId((Long) v));
         });
 
         return modelMapper;
